Adding generic improvements
