--- src/main/java/com/bcnc/exerciseOne/ExerciseOneApplication.java
@@ -1,13 +1,0 @@
-package com.bcnc.exerciseOne;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ExerciseOneApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ExerciseOneApplication.class, args);
-	}
-
-}
