--- src/main/java/com/bcnc/exerciseOne/bridge/AlbumBridge.java
@@ -1,10 +1,0 @@
-package com.bcnc.exerciseOne.bridge;
-
-import com.bcnc.exerciseOne.service.dto.AlbumDto;
-
-import java.util.List;
-
-
-public interface AlbumBridge {
-    List<AlbumDto> getList();
-}
