--- src/main/java/com/bcnc/exerciseOne/bridge/PhotoBridge.java
@@ -1,10 +1,0 @@
-package com.bcnc.exerciseOne.bridge;
-
-import com.bcnc.exerciseOne.service.dto.PhotoDto;
-
-import java.util.List;
-
-public interface PhotoBridge {
-
-    List<PhotoDto> getPhotos();
-}
