--- src/main/java/com/bcnc/exerciseOne/entities/Photo.java
@@ -1,75 +1,0 @@
-package com.bcnc.exerciseOne.entities;
-
-import jakarta.persistence.*;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-@Entity
-@Table(name="tb_photo")
-public class Photo implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    @Id
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    private int albumId;
-    private String title;
-    private String url;
-    private String thumbnailUrl;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public int getAlbumId() {
-        return albumId;
-    }
-
-    public void setAlbumId(int albumId) {
-        this.albumId = albumId;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getThumbnailUrl() {
-        return thumbnailUrl;
-    }
-
-    public void setThumbnailUrl(String thumbnailUrl) {
-        this.thumbnailUrl = thumbnailUrl;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Photo photo = (Photo) o;
-        return id == photo.id && albumId == photo.albumId && Objects.equals(title, photo.title) && Objects.equals(url, photo.url) && Objects.equals(thumbnailUrl, photo.thumbnailUrl);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, albumId, title, url, thumbnailUrl);
-    }
-}
