--- src/main/java/com/bcnc/exerciseOne/repository/AlbumRepository.java
@@ -1,9 +1,0 @@
-package com.bcnc.exerciseOne.repository;
-
-import com.bcnc.exerciseOne.entities.Album;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AlbumRepository extends JpaRepository<Album,Long>{
-}
