--- src/main/java/com/bcnc/exerciseOne/service/AlbumsService.java
@@ -1,13 +1,0 @@
-package com.bcnc.exerciseOne.service;
-
-import com.bcnc.exerciseOne.entities.Album;
-
-import java.util.List;
-
-public interface AlbumsService {
-    List<Album> saveAllRemote();
-
-    List<Album> getRemoteAlbums();
-
-    List<Album> getAllFromDB();
-}
