--- src/main/java/com/bcnc/exerciseOne/service/PhotosService.java
@@ -1,13 +1,0 @@
-package com.bcnc.exerciseOne.service;
-
-import com.bcnc.exerciseOne.entities.Photo;
-
-import java.util.List;
-
-public interface PhotosService {
-    List<Photo> getRemotePhotos();
-
-    List<Photo> getAllFromDB();
-
-    List<Photo> saveAllRemote();
-}
