+++ src/main/java/com/bcnc/one/ExerciseOneApplication.java
@@ -1,0 +1,13 @@
+package com.bcnc.one;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ExerciseOneApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ExerciseOneApplication.class, args);
+	}
+
+}
