+++ src/main/java/com/bcnc/one/bridge/AlbumBridge.java
@@ -1,0 +1,11 @@
+package com.bcnc.one.bridge;
+
+import com.bcnc.one.service.dto.AlbumDto;
+
+import java.util.LinkedHashMap;
+import java.util.List;
+
+
+public interface AlbumBridge {
+    List<LinkedHashMap> getList();
+}
