+++ src/main/java/com/bcnc/one/entities/Album.java
@@ -1,0 +1,65 @@
+package com.bcnc.one.entities;
+
+import jakarta.persistence.*;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+@Entity
+@Table(name="tb_album")
+public class Album implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    @Column(name = "id", nullable = false)
+    private Long id;
+    private int userId;
+
+    private String title;
+
+    @OneToMany(mappedBy = "albumId")
+    private List<Photo> photos = new ArrayList<>();
+
+    public int getUserId() {
+        return userId;
+    }
+
+    public void setUserId(int userId) {
+        this.userId = userId;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public List<Photo> getPhotos() {
+        return photos;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Album album = (Album) o;
+        return userId == album.userId && Objects.equals(id, album.id) && Objects.equals(title, album.title) && Objects.equals(photos, album.photos);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, userId, title, photos);
+    }
+}
