+++ src/main/java/com/bcnc/one/repository/AlbumRepository.java
@@ -1,0 +1,9 @@
+package com.bcnc.one.repository;
+
+import com.bcnc.one.entities.Album;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AlbumRepository extends JpaRepository<Album,Long>{
+}
