+++ src/main/java/com/bcnc/one/repository/PhotoRepository.java
@@ -1,0 +1,7 @@
+package com.bcnc.one.repository;
+
+import com.bcnc.one.entities.Photo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PhotoRepository extends JpaRepository<Photo, Long> {
+}
