+++ src/main/java/com/bcnc/one/service/AlbumsService.java
@@ -1,0 +1,13 @@
+package com.bcnc.one.service;
+
+import com.bcnc.one.entities.Album;
+
+import java.util.List;
+
+public interface AlbumsService {
+    List<Album> saveAllRemote();
+
+    List<Album> getRemoteAlbums();
+
+    List<Album> getAllFromDB();
+}
