+++ src/main/java/com/bcnc/one/service/dto/AlbumDto.java
@@ -1,0 +1,34 @@
+package com.bcnc.one.service.dto;
+
+public class AlbumDto {
+
+    private Long id;
+    private int userId;
+    private String title;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public int getUserId() {
+        return userId;
+    }
+
+    public void setUserId(int userId) {
+        this.userId = userId;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+
+}
