--- src/test/java/com/bcnc/exerciseOne/service/AlbumServiceTest.java
@@ -1,34 +1,0 @@
-package com.bcnc.exerciseOne.service;
-
-import org.junit.jupiter.api.*;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-
-@SpringBootTest
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class AlbumServiceTest {
-
-    @Autowired
-    private AlbumsService albumsService;
-
-    @Test
-    @Order(1)
-    void should_getRemoteEnchanced(){
-        assertNotEquals(0, albumsService.getRemoteAlbums().size());
-    }
-
-    @Test
-    @Order(2)
-    void should_saveInDB(){
-        assertNotEquals(0,albumsService.saveAllRemote().size());
-    }
-
-    @Test
-    @Order(2)
-    void should_getAlbumsEnhancedFromDB(){
-        Assertions.assertFalse(albumsService.getAllFromDB().isEmpty());
-    }
-
-}
